update Results to include the styledDiv
