+++ src/App.tsx
@@ -13,15 +13,13 @@ import Results from './Results';
 import Header from './Header';
 import SubmitButton from './SubmitButton';
 
-import styled, { css } from 'styled-components';
-
 const App: React.FC = () => {
     return (
         <Container maxWidth='md' style={{ textAlign: 'center' }}>
             <ThemeProvider theme={theme}>
                 <Header />
                 <Formik initialValues={initialValues} onSubmit={onSubmit}>
-                    {({ values, handleChange }) => (
+                    {({ values }) => (
                         <Form style={{ marginTop: '50px' }}>
                             <Container>
                                 <FormNumberInput
@@ -38,34 +36,29 @@ const App: React.FC = () => {
                                     name='travel'
                                     label='Tax saver ticket'
                                 />
-
                                 <DynamicField
                                     values={values.bonuses}
                                     addButtonText='shikaka'
                                 />
-
                                 <FormNumberInput
                                     name='property_tax'
                                     label='Property Tax (LPT)'
                                 />
                                 <SubmitButton />
                             </Container>
-                            <StyledDiv standOut>
-                                {/* {JSON.stringify(values, null, 2)} */}
 
-                                {values.salary && values.salary > 0 && (
-                                    <Results
-                                        salary={values.salary / 12}
-                                        bik={values.bonuses.reduce(
-                                            (acc, curr) => curr.value! + acc,
-                                            0
-                                        )}
-                                        travel={values.travel!}
-                                        tax_credit={values.tax_credit}
-                                        property_tax={values.property_tax!}
-                                    />
+                            {/* {JSON.stringify(values, null, 2)} */}
+
+                            <Results
+                                salary={values.salary}
+                                bik={values.bonuses.reduce(
+                                    (acc, curr) => curr.value! + acc,
+                                    0
                                 )}
-                            </StyledDiv>
+                                travel={values.travel!}
+                                tax_credit={values.tax_credit}
+                                property_tax={values.property_tax!}
+                            />
                         </Form>
                     )}
                 </Formik>
@@ -73,18 +66,5 @@ const App: React.FC = () => {
         </Container>
     );
 };
-
-interface divProps {
-    readonly standOut?: boolean;
-}
-
-const StyledDiv = styled.div<divProps>`
-    margin-bottom: 25px;
-    ${(props: divProps) =>
-        props.standOut &&
-        css`
-            margin: 50px;
-        `}
-`;
 
 export default App;
